% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatuR-package.R
\docType{package}
\name{temperatuR-package}
\alias{temperatuR}
\alias{temperatuR-package}
\title{temperatuR: This package is a non-official wrapper for the Temperatur.nu API}
\description{
This package is a non-official wrapper for the Temperatur.nu API. It does not aim to completely map the API, but to wrap some useful pieces of functionality offered by the API and expose them through easy-to-use methods.
}
\details{
This package is a non-official wrapper for the Temperatur.nu API.

More about the API you can find in \url{https://www.temperatur.nu/info/api/}

More about the project you can find in \url{https://www.temperatur.nu/info/}
}
\author{
\strong{Maintainer}: Elisio Quintino \email{elife599@student.liu.se}

Authors:
\itemize{
  \item Carolina Rosario \email{carde734@student.liu.se}
}

}
\keyword{internal}
