\name{temperatuR-package}
\alias{temperatuR-package}
\alias{temperatuR}
\docType{package}
\title{
\packageTitle{temperatuR}
}
\description{
\packageDescription{temperatuR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{temperatuR}
\packageIndices{temperatuR}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{temperatuR}

Maintainer: \packageMaintainer{temperatuR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
